added whole-archive for the client_core library
The operation_rule_execution_manager_factory was not being used by the icommands, so was not being linked in. However, this meant the symbol was missing when the icommands attempted to dynamically load the network plugins, which does use this symbol. To fix this problem, we force the linker to link the whole archive.